More reliable graphing
The setInterval() javascript method is not suitable for web requests, because repeated requests may be make while others are pending.  New code uses setTimeout to do this more reliably.

Also a bug on the server side was fixed.  If data request was made when no new points were available the server would start from the first saved point rather than returning an empty data set, which it does now.