Light control
This enables light control for the first time in CBASS-32.  It is simpler than before.  Keyword LIGHTON followed by a time is when lights turn on.  LIGHTOFF followed by a time turns them off.  Only one of each is supported and ON must be before OFF.  The separate LightPln.ini file is no longer used.

This also fixes a bug in logging the "logLabel" String by converting it to a C string.